PR 24/N (Stage 3): composable audit pass
Applies the criteria agreed at the Stage 3 grill (Q7):
  - Owns reactive state? No → downgrade to utility or factory.
  - Called from one place and < 50 lines? Inline it.
  - Multiple responsibilities? Split.

Three moves:

**Inline `use-translatable-collections-content` into `use-translatable-collections`.**
The former was a 26-line pure wrapper around `useFieldsStore` and `useRelationsStore`,
adapted into a `DirectusDataModel`. Single call site (the latter), so the indirection
wasn't earning its keep. Construct the adapter inline at the one place that needs it.

**Inline `use-sync-container-init` into `Sync.vue`.**
~40 lines, single caller (Sync.vue), pure setup state (refs + one watch). The
historical separation made sense when `useHydrate` was 442 lines and `Sync.vue`
needed help, but now both files are reasonable and the indirection costs more
than it saves.

**Move `use-localazy-singleton.ts` → `stores/singleton-factory.ts`.**
The file's export was `createSingletonStore` — a factory that returns Pinia
setup functions, not a composable in the Vue sense. The misleading filename
+ directory was a holdover from when I first wrote it. Belongs in `stores/`
next to the singleton stores it produces.

After audit the composables directory holds 11 composables, all with either
reactive state, multiple consumers, or store-access wrapping. Each one earns
its `use*` prefix.

Net diff: +46 / -82 across 8 files. 92 tests still passing.
