+++ extensions/module/src/stores/singleton-factory.ts
